Assert Mapping implemented & tested
The following features were implemented:
- Asserting the operations on a mapper
- Full compare of expected and actual mappings
- Detecting multiple asserts of the same operation
- Detecting multiple asserts of the same destination field
- Checking the expected and actual null-strategy on replace operations
- Testing of the transformation function against null input when
strategy is not skip-when-null

The following features were not implemented:
- Testing transformation functions against sample data: We cannot simply
call the transformation function with example data like a recursively
instantiated source object with primitive-default values. We do not know
if that object meets the definition of the transformation method. Plus:
The transformation test should be tested by the user including the happy
path.